Split out C FFI code into separate crate
