+++ .github/workflows/release.yml
@@ -168,14 +168,14 @@ jobs:
           rustup target add x86_64-unknown-linux-musl
           cargo build --target x86_64-unknown-linux-musl --release
           mkdir -p languages/python/native
-          cp polar/polar.h languages/python/native/
+          cp polar-c-api/polar.h languages/python/native/
           cp target/x86_64-unknown-linux-musl/release/libpolar.a languages/python/native/
       - name: Build Rust library (OSX)
         if: runner.os != 'Linux'
         run: |
           cargo build --release
           mkdir -p languages/python/native
-          cp polar/polar.h languages/python/native/
+          cp polar-c-api/polar.h languages/python/native/
           cp target/release/libpolar.a languages/python/native/
 
       - name: Build wheels
