+++ polar-c-api/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "polar-c-api"
+version = "0.1.0"
+authors = ["Oso Security, Inc. <support@osohq.com>"]
+edition = "2018"
+
+[lib]
+name = "polar"
+crate-type = ["lib", "staticlib", "cdylib"]
+bench = false
+
+[dependencies]
+polar-core = { path = "../polar-core" }
+serde = { version = "1.0", features = ["derive", "rc"]}
+serde_json = "1.0"
+
+[build_dependencies]
+cbindgen = "0.14.2"
+
+[features]
+default = []
