+++ polar-c-api/build.rs
@@ -1,12 +1,9 @@
 extern crate cbindgen;
-extern crate lalrpop;
 
 use std::env;
 
 fn main() {
     let crate_dir = env::var("CARGO_MANIFEST_DIR").unwrap();
-
-    lalrpop::process_root().unwrap();
 
     cbindgen::Builder::new()
         .with_crate(crate_dir)
