+++ polar-c-api/src/lib.rs
@@ -1,24 +1,6 @@
-#[macro_use]
-pub mod macros;
-
-#[cfg(test)]
-#[macro_use]
-extern crate maplit;
-
-mod debugger;
-pub mod error;
-mod formatting;
-mod lexer;
-mod numerics;
-pub mod parser;
-mod polar;
-mod rewrites;
-pub mod types;
-mod vm;
-
-pub use self::polar::{Polar, Query};
-pub use error::{PolarError, PolarResult};
-pub use formatting::{draw, ToPolarString};
+use polar_core::error;
+use polar_core::polar::{Polar, Query};
+use polar_core::types;
 
 use std::cell::RefCell;
 use std::ffi::{CStr, CString};
