+++ polar-core/Cargo.toml
@@ -1,12 +1,8 @@
 [package]
-name = "polar"
+name = "polar-core"
 version = "0.1.0"
-authors = ["oso <dev@osohq.com>"]
+authors = ["Oso Security, Inc. <support@osohq.com>"]
 edition = "2018"
-
-[lib]
-crate-type = ["lib", "staticlib", "cdylib"]
-bench = false
 
 [[bench]]
 name = "bench"
@@ -21,7 +17,6 @@ serde = { version = "1.0", features = ["derive", "rc"]}
 serde_json = "1.0"
 
 [build_dependencies]
-cbindgen = "0.14.2"
 lalrpop = {version = "0.18.1", features = ["lexer"]}
 
 [dev-dependencies]
