+++ polar-core/src/error.rs
