+++ polar-core/src/formatting.rs
