+++ polar-core/src/lib.rs
@@ -1,0 +1,16 @@
+#[cfg(test)]
+#[macro_use]
+extern crate maplit;
+
+mod debugger;
+pub mod error;
+pub mod formatting;
+mod lexer;
+#[macro_use]
+pub mod macros;
+mod numerics;
+pub mod parser;
+pub mod polar;
+mod rewrites;
+pub mod types;
+mod vm;
