+++ scripts/docker_package.sh
@@ -7,5 +7,5 @@ cd languages/python
 docker pull quay.io/pypa/manylinux1_x86_64
 mkdir -p native
 cp ../../target/x86_64-unknown-linux-musl/release/libpolar.a native/
-cp ../../polar/polar.h native/
+cp ../../polar-c-api/polar.h native/
 docker run -v $(pwd):/io quay.io/pypa/manylinux1_x86_64 io/package.sh
